+++ debezium-connector-mysql/src/main/java/io/debezium/connector/mysql/MySqlSchema.java
@@ -5,6 +5,7 @@
  */
 package io.debezium.connector.mysql;
 
+import java.lang.reflect.InvocationTargetException;
 import java.util.Collection;
 import java.util.Map;
 import java.util.Set;
@@ -98,18 +99,6 @@ public class MySqlSchema {
         this.topicSelector = topicSelector;
         this.tableIdCaseInsensitive = tableIdCaseInsensitive;
 
-        String ddlParsingModeStr = config.getString(MySqlConnectorConfig.DDL_PARSER_MODE);
-        DdlParsingMode parsingMode = DdlParsingMode.parse(ddlParsingModeStr, MySqlConnectorConfig.DDL_PARSER_MODE.defaultValueAsString());
-
-        try {
-            this.ddlParser = parsingMode.getParserClass().newInstance();
-            this.ddlChanges = this.ddlParser.getDdlChanges();
-        }
-        catch (InstantiationException | IllegalAccessException e) {
-            // ddl parser constructors are not throwing any exceptions, so this should never happen
-            throw new IllegalArgumentException("Unable to create new instance for ddl parser class " + parsingMode.getParserClass().getCanonicalName());
-        }
-
         // Use MySQL-specific converters and schemas for values ...
         String timePrecisionModeStr = config.getString(MySqlConnectorConfig.TIME_PRECISION_MODE);
         TemporalPrecisionMode timePrecisionMode = TemporalPrecisionMode.parse(timePrecisionModeStr);
@@ -122,9 +111,17 @@ public class MySqlSchema {
         MySqlValueConverters valueConverters = new MySqlValueConverters(decimalMode, timePrecisionMode, bigIntUnsignedMode);
         this.schemaBuilder = new TableSchemaBuilder(valueConverters, schemaNameAdjuster, SourceInfo.SCHEMA);
 
-        this.ddlParser = new MySqlDdlParser(false, valueConverters);
-        this.ddlChanges = new DdlChanges(this.ddlParser.terminator());
-        this.ddlParser.addListener(ddlChanges);
+        String ddlParsingModeStr = config.getString(MySqlConnectorConfig.DDL_PARSER_MODE);
+        DdlParsingMode parsingMode = DdlParsingMode.parse(ddlParsingModeStr, MySqlConnectorConfig.DDL_PARSER_MODE.defaultValueAsString());
+
+        try {
+            this.ddlParser = parsingMode.getParserClass().getConstructor(MySqlValueConverters.class).newInstance(valueConverters);
+            this.ddlChanges = this.ddlParser.getDdlChanges();
+        }
+        catch (InstantiationException | IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
+            // ddl parser constructors are not throwing any exceptions, so this should never happen
+            throw new IllegalArgumentException("Unable to create new instance for ddl parser class " + parsingMode.getParserClass().getCanonicalName());
+        }
 
         // Set up the server name and schema prefix ...
         if (serverName != null) serverName = serverName.trim();
